Introduce method snapshot() to trait GuestMemory
If the GuestMemory trait implementation supports memory hotplug, the
underline GuestMemoryRegion array may change dynamically. So take a
snapshot of current guest memory state to get a stable view of the
underline GuestMemoryRegion array.

There are three possible implementation strategies:
1) If the GuestMemory object is static or doesn't support hotplug,
   snapshot returns a reference to the underline object. There's no
   writer at all, so the overhead is trivial.
2) If the GuestMemory object is protected by RwLock, a
   RwLockReadGuard<T> object is returned. Writers will be blocked until
   all guards have been released.
3) If the GuestMemory object is protected by ArcSwap, a
   Guard<'static, Arc<GuestMemoryMmap>> object is returned. It's an RCU
   like lock scheme, so writers won't be blocked but the underline
   resources will be released until all guards have been released.

Strategy 1 should be adopted for VMMs without support of memory hotplug.
And strategy 3 is preferred for VMMs with support of memory hotplug.
Not sure we should try strategy 2.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>