Add a GuestMemory implementation with RCU protection
As discussed in the thread,
https://github.com/rust-vmm/vm-memory/pull/43
ArcSwap is used to provide RCU-like protection for GuestMemoryMmap
to support guest memory hotplug.

Intel Cloud Hypervisor has an implementation based on ArcSwap to support
memory hotplug:
https://github.com/cloud-hypervisor/cloud-hypervisor/pull/572
But the implementation touches too much files and makes it's hard to
reuse code. So this patch extends the vm-memory crate to support both
memory hotplug and code reuse.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>