guest_memory: introduce GuestMemoryMut
If the GuestMemory trait implementation supports memory hotplug, the
underlying GuestMemoryRegion array may change dynamically; accesses
to memory will take a snapshot of current guest memory map to get
a stable view of the underline GuestMemoryRegion array.  This however
adds a new, major restriction on GuestMemory, namely that it must not
"cheat" and use interior mutability.  Otherwise the snapshot would not
be stable.

However, the idea behind vm-memory support for memory hotplug is exactly to
have a sanctioned way to do interior mutability.  This patch therefore
starts the process by splitting the GuestMemory in two:

- the GuestMemoryMut trait retains the possibility to have interior
  mutability.  It therefore drops the num_regions() and find_region()
  methods, because they return values that can change at any time
  (the borrow checker does actually complain about this if you try
  to implement find_region() in GuestMemoryAtomic).  As a result
  GuestMemoryMut also drops Bytes<GuestAddress>; it will only be
  usable through the memory_map method introduced in the next patch.

- the GuestMemory trait keeps find_region() and Bytes<GuestAddress>,
  so it is backwards compatible.

It's up to the GuestMemory implementor to pick the right trait, i.e.
stick to GuestMemoryMut if they use interior mutability.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>