guest_memory: introduce GuestAddressSpace and GuestMemoryBackend
This provides a way for devices to work on a memory map, independent on
how the whole program stores it.  By using GuestAddressSpace
instead of something like &'a M or Arc<M>, they can access memory even
if the VMM chooses to use a GuestMemoryMut for it.  It is up to the
implementation of GuestMemoryMut to provide a suitable mechanism for
obtaining either a GuestMemory or some smart pointer that dereferences
to GuestMemory.  Most notably, the memory_map can be any guard object,
for example a lock guard such as a RwLockReadGuard.  This would not be
very scalable but would be very simple.

If the VMM instead uses an immutable memory map, i.e. a GuestMemory,
vm-memory already provides trivial implementations of the trait that works
for both references and reference-counted objects.  For references,
the implementation imposes no abstraction cost on device that use
GuestAddressSpace<M> instead of &'a M.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>